[#256] Remove all empty values from descriptor
We define empty as `undefined`, empty string, or an empty array. This avoids we
adding values to the descriptor only because the user focused on an input,
without actually adding anything, which can cause the resulting data package to
be invalid (e.g. by having a resource with `encoding == ''`)

Fixes #256